remove obsolete module identifier code
